server/requirements.txt
alembic==1.13.2
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
mypy==1.11.1
mypy-extensions==1.0.0
pathlib==1.0.1
pyasn1==0.6.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
ruff==0.5.6
six==1.16.0
SQLAlchemy==2.0.32
tomli==2.0.1
types-cryptography==3.3.23.2
types-Flask-Cors==4.0.0.20240806
types-jwt==0.1.3
types-pyasn1==0.6.0.20240402
types-PyJWT==1.7.1
types-python-jose==3.3.4.20240106
types-six==1.16.21.20240513
typing_extensions==4.12.2
Werkzeug==3.0.3
